Vue router, random url generator
