+++ vue-smash/package.json
@@ -11,7 +11,9 @@
     "foundation-sites": "^6.5.3",
     "lodash": "^4.17.11",
     "motion-ui": "^2.0.3",
+    "unique-names-generator": "^2.0.1",
     "vue": "^2.6.6",
+    "vue-router": "^3.0.2",
     "vue-socket.io": "^3.0.7"
   },
   "devDependencies": {
